// extensions/cart-transformer-ts/generated/api.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date string.
   * For example, September 7, 2019 is represented as `"2019-07-16"`.
   */
  Date: { input: any; output: any; }
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date and time string.
   * For example, 3:50 pm on September 7, 2019 in the time zone of UTC (Coordinated Universal Time) is
   * represented as `"2019-09-07T15:50:00Z`".
   */
  DateTime: { input: any; output: any; }
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the date and time but not the timezone which is determined from context.
   *
   * For example, "2018-01-01T00:00:00".
   */
  DateTimeWithoutTimezone: { input: any; output: any; }
  /**
   * A signed decimal number, which supports arbitrary precision and is serialized as a string.
   *
   * Example values: `"29.99"`, `"29.999"`.
   */
  Decimal: { input: any; output: any; }
  /**
   * A function-scoped handle to a refer a resource.
   * The Handle type appears in a JSON response as a String, but it is not intended to be human-readable.
   * Example value: `"10079785100"`
   */
  Handle: { input: any; output: any; }
  /**
   * A [JSON](https://www.json.org/json-en.html) object.
   *
   * Example value:
   * `{
   *   "product": {
   *     "id": "gid://shopify/Product/1346443542550",
   *     "title": "White T-shirt",
   *     "options": [{
   *       "name": "Size",
   *       "values": ["M", "L"]
   *     }]
   *   }
   * }`
   */
  JSON: { input: any; output: any; }
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the time but not the date or timezone which is determined from context.
   * For example, "05:43:21".
   */
  TimeWithoutTimezone: { input: any; output: any; }
  /**
   * Represents an [RFC 3986](https://datatracker.ietf.org/doc/html/rfc3986) and
   * [RFC 3987](https://datatracker.ietf.org/doc/html/rfc3987)-compliant URI string.
   *
   * For example, `"https://example.myshopify.com"` is a valid URL. It includes a scheme (`https`) and a host
   * (`example.myshopify.com`).
   */
  URL: { input: any; output: any; }
  /** A void type that can be used to return a null value from a mutation. */
  Void: { input: any; output: any; }
};

/** Represents a generic custom attribute, such as whether an order is a customer's first. */
export type Attribute = {
  __typename?: 'Attribute';
  /** The key or name of the attribute. For example, `"customersFirstOrder"`. */
  key: Scalars['String']['output'];
  /** The value of the attribute. For example, `"true"`. */
  value?: Maybe<Scalars['String']['output']>;
};

/** Key-value pairs that contains additional information. */
export type AttributeOutput = {
  /** Key or name of the attribute. */
  key: Scalars['String']['input'];
  /** Value of the attribute. */
  value: Scalars['String']['input'];
};
// ...
export type Cart = {
  __typename?: 'Cart';
  /** The attributes associated with the cart. Attributes are represented as key-value pairs. */
  attribute?: Maybe<Attribute>;
  /** Information about the buyer that is interacting with the cart. */
  buyerIdentity?: Maybe<BuyerIdentity>;
  /** A list of lines containing information about the items the customer intends to purchase. */
  lines: Array<CartLine>;
};


/** A cart represents the merchandise that a buyer intends to purchase, and the cost associated with the cart. */
export type CartAttributeArgs = {
  key?: InputMaybe<Scalars['String']['input']>;
};

/** Represents information about the merchandise in the cart. */
export type CartLine = {
  __typename?: 'CartLine';
  /**
   * Retrieve a cart line attribute by key.
   *
   * Cart line attributes are also known as line item properties in Liquid.
   */
  attribute?: Maybe<Attribute>;
  /** The cost of the merchandise line that the buyer will pay at checkout. */
  cost: CartLineCost;
  /** The ID of the cart line. */
  id: Scalars['ID']['output'];
  /** The merchandise that the buyer intends to purchase. */
  merchandise: Merchandise;
  /** The quantity of the merchandise that the customer intends to purchase. */
  quantity: Scalars['Int']['output'];
  /**
   * The selling plan associated with the cart line and the effect that each
   * selling plan has on variants when they're purchased.
   */
  sellingPlanAllocation?: Maybe<SellingPlanAllocation>;
};


/** Represents information about the merchandise in the cart. */
export type CartLineAttributeArgs = {
  key?: InputMaybe<Scalars['String']['input']>;
};

/** The cost of the merchandise line that the buyer will pay at checkout. */
export type CartLineCost = {
  __typename?: 'CartLineCost';
  /** The amount of the merchandise line. */
  amountPerQuantity: MoneyV2;
  /** The compare at amount of the merchandise line. */
  compareAtAmountPerQuantity?: Maybe<MoneyV2>;
  /** The cost of the merchandise line before line-level discounts. */
  subtotalAmount: MoneyV2;
  /** The total cost of the merchandise line. */
  totalAmount: MoneyV2;
};

export type CartLineInput = {
  /** The ID of the cart line. */
  cartLineId: Scalars['ID']['input'];
  /** The quantity of the cart line to be merged.The max quantity is 2000. */
  quantity: Scalars['Int']['input'];
};

/** An operation to apply to the Cart. */
export type CartOperation =
  /** A cart line expand operation. */
  { expand: ExpandOperation; merge?: never; update?: never; }
  |  /** A cart line merge operation. */
  { expand?: never; merge: MergeOperation; update?: never; }
  |  /** A cart line update operation. Only stores on the Shopify Plus plan can use apps with update operations. */
  { expand?: never; merge?: never; update: UpdateOperation; };
// ...
export type ExpandOperation = {
  /** The cart line id to expand. */
  cartLineId: Scalars['ID']['input'];
  /** The cart items to expand. */
  expandedCartItems: Array<ExpandedItem>;
  /** The image of the group. */
  image?: InputMaybe<ImageInput>;
  /** The price adjustment to the group. */
  price?: InputMaybe<PriceAdjustment>;
  /** Title override. If title is not provided, variant title is used. */
  title?: InputMaybe<Scalars['String']['input']>;
};

export type ExpandedItem = {
  /** The CartLine attributes to be added to component line. */
  attributes?: InputMaybe<Array<AttributeOutput>>;
  /** The merchandise id of the expanded item. */
  merchandiseId: Scalars['ID']['input'];
  /** The price adjustment for the expanded item. */
  price?: InputMaybe<ExpandedItemPriceAdjustment>;
  /** The quantity of the expanded item.The max quantity is 2000. */
  quantity: Scalars['Int']['input'];
};

/** A fixed price per unit adjustment to apply to the expanded item. */
export type ExpandedItemFixedPricePerUnitAdjustment = {
  /** The fixed price amount per quantity of the expanded item in presentment currency. */
  amount: Scalars['Decimal']['input'];
};

/** A price adjustment to apply to the expanded item. */
export type ExpandedItemPriceAdjustment = {
  /** The price adjustment for the expanded item. */
  adjustment: ExpandedItemPriceAdjustmentValue;
};
// ...
export type FunctionRunResult = {
  /** Cart operations to run on Cart. */
  operations: Array<CartOperation>;
};

/** Represents information about the metafields associated to the specified resource. */
export type HasMetafields = {
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
};
// ...
export type ImageInput = {
  /** The URL of the image. */
  url: Scalars['URL']['input'];
};

export type Input = {
  __typename?: 'Input';
  /** The cart. */
  cart: Cart;
  /** The CartTransform containing the function. */
  cartTransform: CartTransform;
  /** The localization of the Function execution context. */
  localization: Localization;
  /** The conversion rate between the shop's currency and the currency of the cart. */
  presentmentCurrencyRate: Scalars['Decimal']['output'];
  /** Information about the shop. */
  shop: Shop;
};
// ...
export type Merchandise = CustomProduct | ProductVariant;

/** A cart line merge operation. */
export type MergeOperation = {
  /** The CartLine attributes to be added to the parent line. */
  attributes?: InputMaybe<Array<AttributeOutput>>;
  /** The list of cart lines to merge. */
  cartLines: Array<CartLineInput>;
  /** The image of the group. */
  image?: InputMaybe<ImageInput>;
  /** The product variant that models the group of lines. */
  parentVariantId: Scalars['ID']['input'];
  /** The price adjustment to the group. */
  price?: InputMaybe<PriceAdjustment>;
  /** The name of the group of lines to merge. If it isn't specified, it will use the parent_variant' name. */
  title?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type Metafield = {
  __typename?: 'Metafield';
  /** The data stored in the metafield in JSON format. */
  jsonValue: Scalars['JSON']['output'];
  /**
   * The type of data that the metafield stores in the `value` field.
   * Refer to the list of [supported types](https://shopify.dev/apps/metafields/types).
   */
  type: Scalars['String']['output'];
  /** The data stored in the metafield. Always stored as a string, regardless of the metafield's type. */
  value: Scalars['String']['output'];
};

/** A monetary value with currency. */
export type MoneyV2 = {
  __typename?: 'MoneyV2';
  /** Decimal money amount. */
  amount: Scalars['Decimal']['output'];
  /** Currency of the money. */
  currencyCode: CurrencyCode;
};
// ...
export type PriceAdjustment = {
  /** The percentage price decrease of the price adjustment. */
  percentageDecrease?: InputMaybe<PriceAdjustmentValue>;
};

export type PriceAdjustmentValue = {
  /** The value of the price adjustment. */
  value: Scalars['Decimal']['input'];
};

/** Represents a product. */
export type Product = HasMetafields & {
  __typename?: 'Product';
  /** A unique human-friendly string of the product's title. */
  handle: Scalars['Handle']['output'];
  /** Whether the product has any of the given tags. */
  hasAnyTag: Scalars['Boolean']['output'];
  /** Whether the product has the given tags. */
  hasTags: Array<HasTagResponse>;
  /** A globally-unique identifier. */
  id: Scalars['ID']['output'];
  /** Whether the product is in any of the given collections. */
  inAnyCollection: Scalars['Boolean']['output'];
  /** Whether the product is in the given collections. */
  inCollections: Array<CollectionMembership>;
  /** Whether the product is a gift card. */
  isGiftCard: Scalars['Boolean']['output'];
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
  /** The product type specified by the merchant. */
  productType?: Maybe<Scalars['String']['output']>;
  /** The localized title of the product in the customer’s locale. */
  title: Scalars['String']['output'];
  /** The name of the product's vendor. */
  vendor?: Maybe<Scalars['String']['output']>;
};
// ...
export type ProductVariant = HasMetafields & {
  __typename?: 'ProductVariant';
  /** A globally-unique identifier. */
  id: Scalars['ID']['output'];
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
  /** The product that this variant belongs to. */
  product: Product;
  /** Whether the merchandise requires shipping. */
  requiresShipping: Scalars['Boolean']['output'];
  /** An identifier for the product variant in the shop. Required in order to connect to a fulfillment service. */
  sku?: Maybe<Scalars['String']['output']>;
  /** The localized title of the product variant in the customer’s locale. */
  title?: Maybe<Scalars['String']['output']>;
  /** The weight of the product variant in the unit system specified with `weight_unit`. */
  weight?: Maybe<Scalars['Float']['output']>;
  /** Unit of measurement for weight. */
  weightUnit: WeightUnit;
};
// ...
export type UpdateOperation = {
  /** The ID of the cart line. */
  cartLineId: Scalars['ID']['input'];
  /** The image override for the cart line item. */
  image?: InputMaybe<ImageInput>;
  /** The price adjustment for the cart line item. */
  price?: InputMaybe<UpdateOperationPriceAdjustment>;
  /** The title override for the cart line item. */
  title?: InputMaybe<Scalars['String']['input']>;
};
// ...
export enum WeightUnit {
  /** Metric system unit of mass. */
  Grams = 'GRAMS',
  /** 1 kilogram equals 1000 grams. */
  Kilograms = 'KILOGRAMS',
  /** Imperial system unit of mass. */
  Ounces = 'OUNCES',
  /** 1 pound equals 16 ounces. */
  Pounds = 'POUNDS'
}

export type RunInputVariables = Exact<{ [key: string]: never; }>;


export type RunInput = { __typename?: 'Input', cart: { __typename?: 'Cart', plan?: { __typename?: 'Attribute', value?: string | null } | null, lines: Array<{ __typename?: 'CartLine', id: string, quantity: number, cost: { __typename?: 'CartLineCost', subtotalAmount: { __typename?: 'MoneyV2', amount: any } }, plan?: { __typename?: 'Attribute', value?: string | null } | null }> } };